pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md2pdf"
version = "0.1.0"
description = "Convert Markdown files to PDF with customizable styling"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Right Tech Soft LLC" }
]
keywords = ["markdown", "pdf", "converter", "xhtml2pdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "click>=8.1.0",
    "xhtml2pdf>=0.2.11",
    "markdown>=3.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
md2pdf = "md2pdf.cli:main"

[project.urls]
Homepage = "https://github.com/righttechsoft/md2pdf"
Repository = "https://github.com/righttechsoft/md2pdf"

[tool.hatch.build.targets.wheel]
packages = ["src/md2pdf"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
